+++ android/expoview/src/main/java/host/exp/exponent/notifications/schedulers/Scheduler.java
@@ -1,87 +1,24 @@
 package host.exp.exponent.notifications.schedulers;
 
 import android.content.Context;
-
-import java.util.HashMap;
-
-import host.exp.exponent.notifications.ExponentNotificationManager;
 import host.exp.exponent.notifications.exceptions.UnableToScheduleException;
-import host.exp.exponent.notifications.interfaces.SchedulerInterface;
-import host.exp.exponent.notifications.interfaces.SchedulerModelInterface;
 
-public class Scheduler implements SchedulerInterface {
+public interface Scheduler {
 
-  private SchedulerModelInterface mSchedulerModel;
+  void schedule(String action) throws UnableToScheduleException;
 
-  private Context mApplicationContext;
+  String getIdAsString();
 
-  public Scheduler(SchedulerModelInterface schedulerModel) {
-    this.mSchedulerModel = schedulerModel;
-  }
+  String getOwnerExperienceId();
 
-  @Override
-  public void schedule(String action) throws UnableToScheduleException {
-    if (!mSchedulerModel.shouldBeTriggeredByAction(action)) {
-      return;
-    }
-    long nextAppearanceTime = 0;
+  void cancel();
 
-    try {
-      nextAppearanceTime = mSchedulerModel.getNextAppearanceTime();
-    } catch (IllegalArgumentException e) {
-      throw new UnableToScheduleException();
-    }
+  boolean canBeRescheduled();
 
-    String experienceId = mSchedulerModel.getOwnerExperienceId();
-    int notificationId = mSchedulerModel.getNotificationId();
-    HashMap<String, Object> details = mSchedulerModel.getDetails();
+  String saveAndGetId();
 
-    try {
-      getManager().schedule(experienceId, notificationId, details, nextAppearanceTime, null);
-    } catch (ClassNotFoundException e) {
-      e.printStackTrace();
-    }
-  }
+  void setApplicationContext(Context context);
 
-  @Override
-  public String getIdAsString() {
-    return mSchedulerModel.getIdAsString();
-  }
+  void remove();
 
-  @Override
-  public String getOwnerExperienceId() {
-    return mSchedulerModel.getOwnerExperienceId();
-  }
-
-  @Override
-  public void cancel() {
-    String experienceId = mSchedulerModel.getOwnerExperienceId();
-    int notificationId = mSchedulerModel.getNotificationId();
-    getManager().cancel(experienceId, notificationId);
-  }
-
-  @Override
-  public boolean canBeRescheduled() {
-    return mSchedulerModel.canBeRescheduled();
-  }
-
-  @Override
-  public String saveAndGetId() {
-    return mSchedulerModel.saveAndGetId();
-  }
-
-  @Override
-  public void setApplicationContext(Context context) {
-    mApplicationContext = context.getApplicationContext();
-  }
-
-  @Override
-  public void remove() {
-    cancel();
-    mSchedulerModel.remove();
-  }
-
-  private ExponentNotificationManager getManager() {
-    return new ExponentNotificationManager(mApplicationContext);
-  }
 }
